--- app/src/main/java/com/teamtriad/forpets/util/DateExtensions.kt
@@ -1,14 +1,0 @@
-package com.teamtriad.forpets.util
-
-import java.text.SimpleDateFormat
-import java.util.Locale
-
-fun Long.formatDate(): String {
-    val dateFormat = SimpleDateFormat("MM/dd", Locale.getDefault())
-    return dateFormat.format(this)
-}
-
-fun Long.formatDateWithYear(): String {
-    val dateFormat = SimpleDateFormat("yy/MM/dd", Locale.getDefault())
-    return dateFormat.format(this)
-}
